perf(nuxt): smaller package
refactor: move ComponentNames & DirectiveNames out of componentprops

chore: build composable exports copy of composable names (for nuxt)

chore: try moving the floating-ui into dependency... for now
