+++ packages/nuxt/package.json
@@ -33,13 +33,13 @@
     "bootstrap-vue-next": "workspace:^"
   },
   "dependencies": {
-    "@nuxt/kit": "^3.11.2"
+    "@nuxt/kit": "^3.12.3"
   },
   "devDependencies": {
     "@nuxt/devtools": "^1.3.1",
     "@nuxt/eslint-config": "^0.3.13",
     "@nuxt/module-builder": "^0.6.0",
-    "@nuxt/schema": "^3.11.2",
+    "@nuxt/schema": "^3.12.3",
     "@nuxt/test-utils": "^3.13.0",
     "@rushstack/eslint-patch": "^1.10.3",
     "@types/node": "^20.12.12",
@@ -47,7 +47,7 @@
     "bootstrap-vue-next": "workspace:^",
     "eslint": "^8.57.0",
     "eslint-define-config": "^2.1.0",
-    "nuxt": "^3.11.2",
+    "nuxt": "^3.12.3",
     "prettier": "^3.2.5",
     "unimport": "^3.7.1",
     "vue": "^3.4.27"
